% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurveFitting.R
\name{findSigInterval}
\alias{findSigInterval}
\title{Find Significant Time Intervals}
\usage{
findSigInterval(adjusted_pvalue, threshold = 0.05)
}
\arguments{
\item{adjusted_pvalue}{vector of adjested p-value}

\item{threshold}{p_value cut off}
}
\value{
returns a list of the start and end points of all significant time intervals
}
\description{
Identify the significant time intervals
}
\examples{
p=c(0.04, 0.01, 0.02, 0.04, 0.06, 0.2, 0.06, 0.04)
findSigInterval(p, threshold = 0.05)
}
\references{
Ahmed Metwally (ametwa2@uic.edu)
}
