% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{normalize}
\alias{normalize}
\title{Normalize count matrix}
\usage{
normalize(count, method = "css")
}
\arguments{
\item{count}{count matrix}

\item{method}{normalization method}
}
\description{
Normalize count matrix
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
