% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{visualizeArea}
\alias{visualizeArea}
\title{Visualize significant time interval}
\usage{
visualizeArea(aggretage.df, model_ss, method, start, end, text, group_levels)
}
\arguments{
\item{aggretage.df}{Dataframe has the Count , Group, ID, Time}

\item{model_ss}{The fitted model}

\item{method}{Fitting method (negative binomial or LOWESS)}

\item{start}{Vector of the start points of the time intervals}

\item{end}{Vector of the end points of the time intervals}

\item{text}{Feature name}

\item{group_levels}{Level's name}
}
\description{
It highlights the significant tiem intervals
}
\references{
Ahmed Metwally (ametwa2@uic.edu)
}
