% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{visualizeARHistogram}
\alias{visualizeARHistogram}
\title{Visualize histogram of the Area Ratio}
\usage{
visualizeARHistogram(permuted, text)
}
\arguments{
\item{permuted}{Permutation of the permuted data}

\item{text}{Feature name}
}
\description{
Visualize histogram of the Area Ratio emprical distrbition of one time interval
}
\references{
Ahmed Metwally (ametwa2@uic.edu)
}
