% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{visualizeFeature}
\alias{visualizeFeature}
\title{Visualize Longitudinal Feature}
\usage{
visualizeFeature(df, text, group.levels, unit = "days")
}
\arguments{
\item{df}{dataframe has the Count, Group, ID, Time}

\item{text}{feature name}

\item{group.levels}{The two level's name}

\item{unit}{time interval unit}
}
\description{
Visualize Longitudinal Feature
}
\examples{
data(metalonda_test_data)
n.sample = 5
n.timepoints = 10
n.group = 2
Group = factor(c(rep(0, n.sample*n.timepoints), rep(1, n.sample*n.timepoints)))
Time = rep(rep(1:n.timepoints, times = n.sample), 2)
ID = factor(rep(1:(2*n.sample), each = n.timepoints))
points = seq(1, 10, length.out = 10)
aggregate.df = data.frame(Count = metalonda_test_data[1,], Time = Time, Group = Group, ID = ID)
visualizeFeature(aggregate.df, text = rownames(metalonda_test_data)[1], Group)
}
\references{
Ahmed Metwally (ametwa2@uic.edu)
}
