% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{visualizeFeatureSpline}
\alias{visualizeFeatureSpline}
\title{Visualize Feature with the fitted Splines}
\usage{
visualizeFeatureSpline(df, model, method, text, group_levels)
}
\arguments{
\item{df}{dataframe has the Count , Group, ID, Time}

\item{model}{the fitted model}

\item{method}{The fitting method (negative binomial, LOWESS)}

\item{text}{feature name}

\item{group_levels}{The two levels name}
}
\description{
Plot the longitudinal features along with the fitted splines
}
\references{
Ahmed Metwally (ametwa2@uic.edu)
}
