% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{visualizeVolcanoPlot}
\alias{visualizeVolcanoPlot}
\title{Visualize log2 fold-change and significance of each interval as volcano plot}
\usage{
visualizeVolcanoPlot(df, text)
}
\arguments{
\item{df}{Dataframe has a detailed summary about feature's significant intervals}

\item{text}{Feature name}
}
\description{
Visualize log2 fold-change and significance of each interval as volcano plot
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
