% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{visualizeArea}
\alias{visualizeArea}
\title{Visualize significant time interval}
\usage{
visualizeArea(aggregate.df, model.ss, method, start, end, text, group.levels,
  unit = "days")
}
\arguments{
\item{aggregate.df}{Dataframe has the Count, Group, ID, Time}

\item{model.ss}{The fitted model}

\item{method}{Fitting method (negative binomial or LOWESS)}

\item{start}{Vector of the start points of the time intervals}

\item{end}{Vector of the end points of the time intervals}

\item{text}{Feature name}

\item{group.levels}{Level's name}

\item{unit}{time unit used in the Time vector (days, weeks, months, etc.)}
}
\description{
Visualize significant time interval
}
\references{
Ahmed Metwally (ametwa2@uic.edu)
}

