% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{visualizeARHistogram}
\alias{visualizeARHistogram}
\title{Visualize Area Ratio (AR) empirical distribution}
\usage{
visualizeARHistogram(permuted, text, method)
}
\arguments{
\item{permuted}{Permutation of the permuted data}

\item{text}{Feature name}

\item{method}{fitting method}
}
\description{
Visualize Area Ratio (AR) empirical distribution for each time interval
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
