% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{visualizeFeatureSpline}
\alias{visualizeFeatureSpline}
\title{Visualize the feature trajectory with the fitted Splines}
\usage{
visualizeFeatureSpline(
  df,
  model,
  method,
  text,
  group.levels,
  unit = "days",
  ylabel = "Normalized Count",
  col = c("blue", "firebrick"),
  prefix = "Test"
)
}
\arguments{
\item{df}{dataframe has the Count , Group, ID, Time}

\item{model}{the fitted model}

\item{method}{The fitting method (negative binomial, LOWESS)}

\item{text}{feature name}

\item{group.levels}{The two level's name}

\item{unit}{time unit used in the Time vector (hours, days, weeks, months, etc.)}

\item{ylabel}{text to be shown on the y-axis of all generated figures (default: "Normalized Count")}

\item{col}{two color to be used for the two groups (eg., c("red", "blue")).}

\item{prefix}{prefix to be used to create directory for the analysis results}
}
\description{
Plot the longitudinal features along with the fitted splines
}
\references{
Ahmed Metwally (ametwall@stanford.edu)
}
